feat(compliance): auto-apply high-confidence url-dead recoveries + sync enrichment on insert
