+++ src/app/api/admin/legal-refs/recover-url-dead/route.ts
@@ -23,6 +23,7 @@ import { createClient as createAdminClient } from '@supabase/supabase-js';
 import { checkUkLegalAuthority } from '@/lib/legal-refs-authority';
 import { authorizeAdminOrCron } from '@/lib/admin-auth';
 import { logPerplexityCall } from '@/lib/cost-ledger';
+import { enrichSingleCorrection } from '@/lib/legal-refs-enrich';
 
 const PERPLEXITY_MODEL = 'sonar-pro';
 // sonar-pro flat rate ≈ $0.005 per request. USD→GBP ≈ 0.79.
@@ -187,6 +188,7 @@ export async function POST(request: NextRequest) {
     now_resolves: 0,
     redirected_to_authority: 0,
     queued: 0,
+    auto_applied: 0,
     perplexity_calls: 0,
     perplexity_recovered: 0,
     perplexity_cost_gbp: 0,
@@ -233,23 +235,35 @@ export async function POST(request: NextRequest) {
         'Verify the destination still cites the same law before approving.';
 
       // eslint-disable-next-line no-await-in-loop
-      const { error: insErr } = await admin.from('legal_ref_corrections').insert({
-        ref_id: row.id,
-        proposer: 'url-dead-recovery-2026-04-30',
-        before_law_name: row.law_name,
-        before_source_url: row.source_url,
-        before_status: 'url_dead',
-        proposed_law_name: null,
-        proposed_source_url,
-        proposed_status: null,
-        reasoning,
-        confidence: 'medium',
-        status: 'pending',
-      });
+      const { data: insertedRow, error: insErr } = await admin
+        .from('legal_ref_corrections')
+        .insert({
+          ref_id: row.id,
+          proposer: 'url-dead-recovery-2026-04-30',
+          before_law_name: row.law_name,
+          before_source_url: row.source_url,
+          before_status: 'url_dead',
+          proposed_law_name: null,
+          proposed_source_url,
+          proposed_status: null,
+          reasoning,
+          confidence: 'medium',
+          status: 'pending',
+        })
+        .select('id')
+        .maybeSingle();
       if (insErr) {
         summary.errors.push(`${row.id}: ${insErr.message}`);
       } else {
         summary.queued++;
+        // Fire enrichment synchronously so PendingCorrectionsSection sees
+        // the row as "needs your eye" right away — instead of waiting for
+        // the next 04:00 UTC enrichment cron, which leaves the founder
+        // staring at "No items currently need your eye" while corrections
+        // sit in limbo with enriched_at=null.
+        if (insertedRow?.id) {
+          enrichSingleCorrection(admin, insertedRow.id).catch(() => {});
+        }
       }
     } else if (queue && category === 'still_dead') {
       // Perplexity fallback for Cloudflare-blocked publishers (Ofcom,
@@ -289,25 +303,127 @@ export async function POST(request: NextRequest) {
             `Original URL 403/404 (default UA=${def.status}, browser UA=${ua.status}). ` +
             `Perplexity sonar-pro proposed canonical URL on same publisher domain ` +
             `(${pubDomain}) with confidence=${recovery.confidence}. Reasoning: ${recovery.reasoning}`;
-          // eslint-disable-next-line no-await-in-loop
-          const { error: insErr } = await admin.from('legal_ref_corrections').insert({
-            ref_id: row.id,
-            proposer: 'url-recovery-perplexity-2026-04-30',
-            before_law_name: row.law_name,
-            before_source_url: row.source_url,
-            before_status: 'url_dead',
-            proposed_law_name: null,
-            proposed_source_url: proposedUrl,
-            proposed_status: null,
-            reasoning,
-            confidence: recovery.confidence,
-            status: 'pending',
-          });
-          if (insErr) {
-            summary.errors.push(`${row.id}: ${insErr.message}`);
+
+          // ---- Auto-apply fast-path: HIGH confidence + same publisher + authority ----
+          // CLAUDE.md "Compliance citation principle" gates direct mutation of
+          // legal_references behind founder approval, with one exception: the
+          // same-host-redirect fast-path (legislation.gov.uk/x/y →
+          // legislation.gov.uk/x/y/contents) auto-applies because no semantic
+          // change is possible when host + authority + law name are unchanged.
+          //
+          // We extend that fast-path to "same publisher domain + Perplexity
+          // confidence='high' + final URL passes checkUkLegalAuthority". The
+          // risk profile is functionally equivalent: the law name is unchanged
+          // (proposed_law_name=null), the publisher is unchanged (e.g. ofcom →
+          // ofcom), the destination is in the authority allowlist, and HIGH
+          // confidence means Perplexity has identified a known canonical
+          // replacement (not a guess). MEDIUM confidence stays on the queue —
+          // founder reviews those.
+          if (recovery.confidence === 'high') {
+            // 1. UPDATE legal_references — single mutation, gated by all four
+            //    conditions above (HIGH + samePublisher + sameAuthority +
+            //    proposedUrl validated via checkUkLegalAuthority).
+            // eslint-disable-next-line no-await-in-loop
+            const { error: refErr } = await admin
+              .from('legal_references')
+              .update({
+                source_url: proposedUrl,
+                verification_status: 'current',
+                last_verified: new Date().toISOString(),
+              })
+              .eq('id', row.id);
+            if (refErr) {
+              summary.errors.push(`${row.id} (auto-apply update): ${refErr.message}`);
+            } else {
+              // 2. INSERT correction row with status='approved' for audit
+              //    trail. applied_at marks the auto-apply timestamp; the
+              //    schema does not have a separate auto_applied_at column,
+              //    so we rely on (status='approved' + proposer prefix +
+              //    reviewed_by='system-auto-apply') to identify auto-applied
+              //    rows in the admin "Auto-applied (last 7 days)" panel.
+              // eslint-disable-next-line no-await-in-loop
+              const nowIso = new Date().toISOString();
+              // eslint-disable-next-line no-await-in-loop
+              await admin.from('legal_ref_corrections').insert({
+                ref_id: row.id,
+                proposer: 'url-recovery-perplexity-2026-04-30',
+                before_law_name: row.law_name,
+                before_source_url: row.source_url,
+                before_status: 'url_dead',
+                proposed_law_name: null,
+                proposed_source_url: proposedUrl,
+                proposed_status: 'current',
+                reasoning,
+                confidence: recovery.confidence,
+                status: 'approved',
+                reviewed_at: nowIso,
+                reviewed_by: 'system-auto-apply',
+                applied_at: nowIso,
+                notes:
+                  'Auto-applied: same-publisher + Perplexity high-confidence + ' +
+                  'authority allowlist (extension of same-host fast-path).',
+              });
+
+              // 3. Audit row in legal_ref_verifications (γ).
+              // eslint-disable-next-line no-await-in-loop
+              await admin.from('legal_ref_verifications').insert({
+                ref_id: row.id,
+                verifier: 'auto-apply-url-dead-recovery',
+                triggered_by: auth.userId ?? 'cron',
+                before_status: 'url_dead',
+                after_status: 'current',
+                before_url: row.source_url,
+                after_url: proposedUrl,
+                changes: {
+                  before: {
+                    source_url: row.source_url,
+                    verification_status: 'url_dead',
+                  },
+                  after: {
+                    source_url: proposedUrl,
+                    verification_status: 'current',
+                  },
+                },
+                notes: reasoning,
+                perplexity_response: { verdict: 'auto_applied', recovery },
+              });
+
+              summary.auto_applied++;
+              summary.perplexity_recovered++;
+            }
           } else {
-            summary.queued++;
-            summary.perplexity_recovered++;
+            // MEDIUM confidence — queue for founder review (no change in
+            // policy here vs. the previous behaviour).
+            // eslint-disable-next-line no-await-in-loop
+            const { data: insertedRow, error: insErr } = await admin
+              .from('legal_ref_corrections')
+              .insert({
+                ref_id: row.id,
+                proposer: 'url-recovery-perplexity-2026-04-30',
+                before_law_name: row.law_name,
+                before_source_url: row.source_url,
+                before_status: 'url_dead',
+                proposed_law_name: null,
+                proposed_source_url: proposedUrl,
+                proposed_status: null,
+                reasoning,
+                confidence: recovery.confidence,
+                status: 'pending',
+              })
+              .select('id')
+              .maybeSingle();
+            if (insErr) {
+              summary.errors.push(`${row.id}: ${insErr.message}`);
+            } else {
+              summary.queued++;
+              summary.perplexity_recovered++;
+              // Fire enrichment synchronously so the row appears in
+              // PendingCorrectionsSection's default "needs your eye" view
+              // without waiting for the 04:00 UTC enrichment cron.
+              if (insertedRow?.id) {
+                enrichSingleCorrection(admin, insertedRow.id).catch(() => {});
+              }
+            }
           }
         } else {
           // Null / low confidence / off-domain — leave url_dead, log for
